Move state to it's own file
