+++ device-types/src/lib.rs
@@ -29,7 +29,6 @@ pub mod state {
             }
         }
     }
-
 }
 
 pub mod mount {
